test: Rearranged tests
